fix: 신청서 age 추가

+++ src/main/java/studio/stew/dto/ApplicationResponseDto.java
@@ -130,7 +130,7 @@ public class ApplicationResponseDto {
         String title;
         String userName;
         Gender gender;
-        Integer age;
+        Long age;
         String location;
         List<String> purpose;
         int intensity;
